iaq-agent/requirements.txt
pika==1.3.2
pandas==2.1.4
python-dotenv==1.0.0
structlog==23.2.0
fastapi==0.104.1
uvicorn==0.24.0